+++ apple-bundles/src/directory_bundle.rs
@@ -11,7 +11,7 @@ use {
         collections::HashSet,
         path::{Path, PathBuf},
     },
-    tugger_file_manifest::{is_executable, FileEntry, FileManifest},
+    simple_file_manifest::{is_executable, FileEntry, FileManifest},
 };
 
 /// An Apple bundle backed by a filesystem/directory.