+++ apple-bundles/src/macos_application_bundle.rs
@@ -12,7 +12,7 @@ use {
     crate::BundlePackageType,
     anyhow::{anyhow, Context, Result},
     std::path::{Path, PathBuf},
-    tugger_file_manifest::{FileEntry, FileManifest, FileManifestError},
+    simple_file_manifest::{FileEntry, FileManifest, FileManifestError},
 };
 
 /// Primitive used to iteratively construct a macOS Application Bundle.
@@ -46,7 +46,7 @@ use {
 ///
 /// ```
 /// use apple_bundles::MacOsApplicationBundleBuilder;
-/// use tugger_file_manifest::FileEntry;
+/// use simple_file_manifest::FileEntry;
 ///
 /// # fn main() -> anyhow::Result<()> {
 /// let mut builder = MacOsApplicationBundleBuilder::new("MyProgram")?;